[Silver III] Title: 주유소, Time: 20 ms, Memory: 2800 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -31,7 +31,7 @@ int main() {
 		if (min_v > cost[i]) {
 			min_v = cost[i];
 		}
-		ans += (min_v * road[i]);
+		ans += ((long long)min_v * (long long)road[i]);
 	}
 	
 	cout << ans;
